Updated code as per the review comments
