+++ feature/policy_forwarding/decapsulation/otg_tests/ipv4_guev1_decap_and_hashing_test/ipv4_guev1_decap_and_hashing_test.go
@@ -16,7 +16,6 @@ package ipv4_guev1_decap_and_hashing_test
 
 import (
 	"bytes"
-	"context"
 	"encoding/binary"
 	"fmt"
 	"net"
@@ -31,7 +30,6 @@ import (
 	"github.com/openconfig/featureprofiles/internal/deviations"
 	"github.com/openconfig/featureprofiles/internal/fptest"
 	"github.com/openconfig/featureprofiles/internal/otgutils"
-	gpb "github.com/openconfig/gnmi/proto/gnmi"
 	"github.com/openconfig/ondatra"
 	"github.com/openconfig/ondatra/gnmi"
 	"github.com/openconfig/ondatra/gnmi/oc"
@@ -69,14 +67,14 @@ const (
 	trafficFrameSize     = 1500
 	ratePercent          = 10
 	lspV4Name            = "lsp-egress-v4"
-	rplName              = "ALLOW"
-	mplsLabel            = 1000
-	decapType            = "udp"
-	decapPort            = 6080
-	ecmpMaxPath          = 4
-	policyName           = "decap-policy"
-	policyId             = 1
-	trafficDuration      = 20
+	// rplName              = "ALLOW"
+	mplsLabel       = 1000
+	decapType       = "udp"
+	decapPort       = 6080
+	ecmpMaxPath     = 4
+	policyName      = "decap-policy"
+	policyId        = 1
+	trafficDuration = 20
 )
 
 // IP Addresses and Attributes
@@ -353,7 +351,7 @@ func configureDUT(t *testing.T, dut *ondatra.DUTDevice) []string {
 		dut.Port(t, "port4"),
 	}
 	aggIDsList = append(aggIDsList, aggID1)
-	clearLAGInterfaces(t, dut, dutAggPorts1, aggID1)
+	cfgplugins.DeleteAggregate(t, dut, aggID1, dutAggPorts1)
 	configureDUTLag(t, dut, dutAggPorts1, aggID1, dutLag1)
 
 	// Ports 5 and 6 will be part of LAG
@@ -362,6 +360,7 @@ func configureDUT(t *testing.T, dut *ondatra.DUTDevice) []string {
 		dut.Port(t, "port6"),
 	}
 	aggIDsList = append(aggIDsList, aggID2)
+	cfgplugins.DeleteAggregate(t, dut, aggID2, dutAggPorts2)
 	configureDUTLag(t, dut, dutAggPorts2, aggID2, dutLag2)
 	fptest.ConfigureDefaultNetworkInstance(t, dut)
 
@@ -369,95 +368,22 @@ func configureDUT(t *testing.T, dut *ondatra.DUTDevice) []string {
 	dutConfPath := d.NetworkInstance(deviations.DefaultNetworkInstance(dut)).Protocol(oc.PolicyTypes_INSTALL_PROTOCOL_TYPE_ISIS, isisInstance)
 	dutConf := configureDUTISIS(t, dut, p1, p2, aggID1)
 	gnmi.Replace(t, dut, dutConfPath.Config(), dutConf)
+	cfgplugins.ConfigureDUTBGP(t, dut, dutAS, ate1AS, ate2AS, ate3AS, ate4AS, ate5AS, ecmpMaxPath, dutP1, dutP2, dutP7,
+		ateP1, ateP2, ateP7, dutLag1, dutLag2, ateLag1, ateLag2)
 
-	dutBgpConfPath := d.NetworkInstance(deviations.DefaultNetworkInstance(dut)).Protocol(oc.PolicyTypes_INSTALL_PROTOCOL_TYPE_BGP, "BGP")
-
-	// Create BGP config only once
-	dutBgpConf := &oc.NetworkInstance_Protocol{
-		Identifier: oc.PolicyTypes_INSTALL_PROTOCOL_TYPE_BGP,
-		Name:       ygot.String("BGP"),
-		Bgp:        &oc.NetworkInstance_Protocol_Bgp{},
-	}
-
-	bgp := dutBgpConf.Bgp
-	global := bgp.GetOrCreateGlobal()
-	global.As = ygot.Uint32(dutAS)
-	global.RouterId = ygot.String(dutP1.IPv4)
-
-	af4 := global.GetOrCreateAfiSafi(oc.BgpTypes_AFI_SAFI_TYPE_IPV4_UNICAST)
-	af4.Enabled = ygot.Bool(true)
-	af6 := global.GetOrCreateAfiSafi(oc.BgpTypes_AFI_SAFI_TYPE_IPV6_UNICAST)
-	af6.Enabled = ygot.Bool(true)
-
-	// Append EBGP Neighbors - ATE1 and ATE5
-	appendBGPNeighbor(t, bgp, ate1AS, dutP1.Name, ateP1.IPv4, ateP1.IPv6, false)
-	appendBGPNeighbor(t, bgp, ate5AS, dutP7.Name, ateP7.IPv4, ateP7.IPv6, false)
-	// Append IBGP Neighbor - ATE2
-	appendBGPNeighbor(t, bgp, ate2AS, dutP2.Name, ateP2.IPv4, ateP2.IPv6, false)
-	appendBGPNeighbor(t, bgp, ate3AS, dutLag1.Name, ateLag1.IPv4, ateLag1.IPv6, true)
-	appendBGPNeighbor(t, bgp, ate4AS, dutLag2.Name, ateLag2.IPv4, ateLag2.IPv6, true)
-
-	// Apply the whole BGP config
-	gnmi.Replace(t, dut, dutBgpConfPath.Config(), dutBgpConf)
-	if deviations.LoadIntervalNotSupported(dut) {
-		gnmiClient := dut.RawAPIs().GNMI(t)
-		jsonConfig := `
-		load-balance policies
-		load-balance sand profile default
-		fields ipv6 outer dst-ip flow-label next-header src-ip
-        fields l4 outer dst-port src-port
-        no fields mpls
-        packet-type gue outer-ip
-		`
-		gpbSetRequest := buildCliConfigRequest(jsonConfig)
-
-		if _, err := gnmiClient.Set(context.Background(), gpbSetRequest); err != nil {
-			t.Fatalf("gnmiClient.Set() with unexpected error: %v", err)
-		}
-	} else {
-		// TODO: OC does not yet support selecting the load-balancing hash mode on LAG members.
-		t.Logf("Load balancing is currently not supported via OpenConfig. Will fix once it's implemented.")
-	}
-
-	if deviations.MultipathUnsupportedNeighborOrAfisafi(dut) {
-		t.Log("Executing CLI commands")
-		gnmiClient := dut.RawAPIs().GNMI(t)
-		jsonConfig := fmt.Sprintf(`		
-		router bgp %d
-		address-family ipv4
-		maximum-paths %[2]d ecmp %[2]d
-		bgp bestpath as-path multipath-relax
-		address-family ipv6
-		maximum-paths %[2]d ecmp %[2]d
-		bgp bestpath as-path multipath-relax
-		`, dutAS, ecmpMaxPath)
-		gpbSetRequest := buildCliConfigRequest(jsonConfig)
-
-		if _, err := gnmiClient.Set(context.Background(), gpbSetRequest); err != nil {
-			t.Fatalf("gnmiClient.Set() with unexpected error: %v", err)
-		}
-	} else {
-		// TODO: As per the latest OpenConfig GNMI OC schema — the Encapsulation/Decapsulation sub-tree is not fully implemented, need to add OC commands once implemented.
-		af4.GetOrCreateUseMultiplePaths().Enabled = ygot.Bool(true)
-		af4.GetOrCreateUseMultiplePaths().GetOrCreateEbgp().AllowMultipleAs = ygot.Bool(true)
-		af6.GetOrCreateUseMultiplePaths().Enabled = ygot.Bool(true)
-		af6.GetOrCreateUseMultiplePaths().GetOrCreateEbgp().AllowMultipleAs = ygot.Bool(true)
-		af4.GetOrCreateUseMultiplePaths().GetOrCreateIbgp().SetMaximumPaths(ecmpMaxPath)
-		af6.GetOrCreateUseMultiplePaths().GetOrCreateIbgp().SetMaximumPaths(ecmpMaxPath)
-	}
 	return aggIDsList
 }
 
 func configureDutWithGueDecap(t *testing.T, dut *ondatra.DUTDevice, ipType string) {
 	t.Logf("Configure DUT with decapsulation UDP port %v", decapPort)
-	ocPFParams := GetDefaultOcPolicyForwardingParams(t, dut, ipType)
+	ocPFParams := getDefaultOcPolicyForwardingParams(t, dut, ipType)
 	_, _, pf := cfgplugins.SetupPolicyForwardingInfraOC(ocPFParams.NetworkInstanceName)
 	cfgplugins.DecapGroupConfigGue(t, dut, pf, ocPFParams)
 }
 
-// GetDefaultOcPolicyForwardingParams provides default parameters for the generator,
+// tetDefaultOcPolicyForwardingParams provides default parameters for the generator,
 // matching the values in the provided JSON example.
-func GetDefaultOcPolicyForwardingParams(t *testing.T, dut *ondatra.DUTDevice, ipType string) cfgplugins.OcPolicyForwardingParams {
+func getDefaultOcPolicyForwardingParams(t *testing.T, dut *ondatra.DUTDevice, ipType string) cfgplugins.OcPolicyForwardingParams {
 	return cfgplugins.OcPolicyForwardingParams{
 		NetworkInstanceName: "DEFAULT",
 		InterfaceID:         dut.Port(t, "port1").Name(),
@@ -474,7 +400,7 @@ func configureDUTLag(t *testing.T, dut *ondatra.DUTDevice, aggPorts []*ondatra.P
 	for _, port := range aggPorts {
 		gnmi.Delete(t, dut, gnmi.OC().Interface(port.Name()).Ethernet().Config())
 	}
-	setupAggregateAtomically(t, dut, aggPorts, aggID)
+	cfgplugins.SetupStaticAggregateAtomically(t, dut, aggPorts, aggID)
 	agg := dutLag.NewOCInterface(aggID, dut)
 	agg.Type = oc.IETFInterfaces_InterfaceType_ieee8023adLag
 	agg.GetOrCreateAggregation().LagType = oc.IfAggregate_AggregationType_STATIC
@@ -489,25 +415,6 @@ func configureDUTLag(t *testing.T, dut *ondatra.DUTDevice, aggPorts []*ondatra.P
 		}
 		gnmi.Replace(t, dut, gnmi.OC().Interface(port.Name()).Config(), i)
 	}
-}
-
-func setupAggregateAtomically(t *testing.T, dut *ondatra.DUTDevice, aggPorts []*ondatra.Port, aggID string) {
-	t.Helper()
-	d := &oc.Root{}
-	agg := d.GetOrCreateInterface(aggID)
-	agg.Type = oc.IETFInterfaces_InterfaceType_ieee8023adLag
-	agg.GetOrCreateAggregation().LagType = oc.IfAggregate_AggregationType_STATIC
-
-	for _, port := range aggPorts {
-		i := d.GetOrCreateInterface(port.Name())
-		i.GetOrCreateEthernet().AggregateId = ygot.String(aggID)
-		i.Type = oc.IETFInterfaces_InterfaceType_ethernetCsmacd
-
-		if deviations.InterfaceEnabled(dut) {
-			i.Enabled = ygot.Bool(true)
-		}
-	}
-	gnmi.Update(t, dut, gnmi.OC().Config(), d)
 }
 
 func clearLAGInterfaces(t *testing.T, dut *ondatra.DUTDevice, aggPorts []*ondatra.Port, aggID string) {
@@ -554,52 +461,6 @@ func configureDUTISIS(t *testing.T, dut *ondatra.DUTDevice, p1, p2 *ondatra.Port
 	isisLo.Passive = ygot.Bool(true)
 
 	return protocol
-}
-
-func appendBGPNeighbor(t *testing.T, bgp *oc.NetworkInstance_Protocol_Bgp, ateAs uint32, portName, neighborIpV4, neighborIpV6 string, isLag bool) {
-	t.Helper()
-	// Peer Group for IPv4
-	pgv4 := bgp.GetOrCreatePeerGroup(portName + "BGP-PEER-GROUP-V4")
-	pgv4.PeerAs = ygot.Uint32(ateAs)
-	pgv4.PeerGroupName = ygot.String(portName + "BGP-PEER-GROUP-V4")
-	pgafv4 := pgv4.GetOrCreateAfiSafi(oc.BgpTypes_AFI_SAFI_TYPE_IPV4_UNICAST)
-	pgafv4.Enabled = ygot.Bool(true)
-	rpl4 := pgafv4.GetOrCreateApplyPolicy()
-	rpl4.ImportPolicy = []string{rplName}
-	rpl4.ExportPolicy = []string{rplName}
-
-	// Peer Group for IPv6
-	pgv6 := bgp.GetOrCreatePeerGroup(portName + "BGP-PEER-GROUP-V6")
-	pgv6.PeerAs = ygot.Uint32(ateAs)
-	pgv6.PeerGroupName = ygot.String(portName + "BGP-PEER-GROUP-V6")
-	pgafv6 := pgv6.GetOrCreateAfiSafi(oc.BgpTypes_AFI_SAFI_TYPE_IPV6_UNICAST)
-	pgafv6.Enabled = ygot.Bool(true)
-	rpl6 := pgafv6.GetOrCreateApplyPolicy()
-	rpl6.ImportPolicy = []string{rplName}
-	rpl6.ExportPolicy = []string{rplName}
-
-	// IPv4 Neighbor
-	nv4 := bgp.GetOrCreateNeighbor(neighborIpV4)
-	nv4.PeerAs = ygot.Uint32(ateAs)
-	nv4.Enabled = ygot.Bool(true)
-	nv4.PeerGroup = ygot.String(portName + "BGP-PEER-GROUP-V4")
-	afisafi4 := nv4.GetOrCreateAfiSafi(oc.BgpTypes_AFI_SAFI_TYPE_IPV4_UNICAST)
-	afisafi4.Enabled = ygot.Bool(true)
-	nv4.GetOrCreateAfiSafi(oc.BgpTypes_AFI_SAFI_TYPE_IPV6_UNICAST).Enabled = ygot.Bool(false)
-
-	// IPv6 Neighbor
-	nv6 := bgp.GetOrCreateNeighbor(neighborIpV6)
-	nv6.PeerAs = ygot.Uint32(ateAs)
-	nv6.Enabled = ygot.Bool(true)
-	// Enable multihop on LAGs
-	if isLag {
-		nv4.GetOrCreateEbgpMultihop().SetMultihopTtl(5)
-		nv6.GetOrCreateEbgpMultihop().SetMultihopTtl(5)
-	}
-	nv6.PeerGroup = ygot.String(portName + "BGP-PEER-GROUP-V6")
-	afisafi6 := nv6.GetOrCreateAfiSafi(oc.BgpTypes_AFI_SAFI_TYPE_IPV6_UNICAST)
-	afisafi6.Enabled = ygot.Bool(true)
-	nv6.GetOrCreateAfiSafi(oc.BgpTypes_AFI_SAFI_TYPE_IPV4_UNICAST).Enabled = ygot.Bool(false)
 }
 
 func configureATE(t *testing.T, ate *ondatra.ATEDevice) gosnappi.Config {
@@ -1128,24 +989,4 @@ func verifySinglePathTraffic(t *testing.T, ate *ondatra.ATEDevice, otgConfig gos
 	} else {
 		t.Logf("PASS: All traffic followed a single path as expected")
 	}
-}
-
-// Support method to execute GNMIC commands
-func buildCliConfigRequest(config string) *gpb.SetRequest {
-	gpbSetRequest := &gpb.SetRequest{
-		Update: []*gpb.Update{
-			{
-				Path: &gpb.Path{
-					Origin: "cli",
-					Elem:   []*gpb.PathElem{},
-				},
-				Val: &gpb.TypedValue{
-					Value: &gpb.TypedValue_AsciiVal{
-						AsciiVal: config,
-					},
-				},
-			},
-		},
-	}
-	return gpbSetRequest
 }
